refactor restoreResourceVersionPriority to file
